+++ opcua_plugin/opcua.go
@@ -42,7 +42,8 @@ var OPCUAConfigSpec = service.NewConfigSpec().
 	Field(service.NewBoolField("subscribeEnabled").Description("Set to true to subscribe to OPC UA nodes instead of fetching them every seconds. Default is pulling messages every second (false).").Default(false)).
 	Field(service.NewBoolField("directConnect").Description("Set this to true to directly connect to an OPC UA endpoint. This can be necessary in cases where the OPC UA server does not allow 'endpoint discovery'. This requires having the full endpoint name in endpoint, and securityMode and securityPolicy set. Defaults to 'false'").Default(false)).
 	Field(service.NewBoolField("useHeartbeat").Description("Set to true to provide an extra message with the servers timestamp as a heartbeat").Default(false)).
-	Field(service.NewBoolField("browseHierarchicalReferences").Description("Set to true to browse hierarchical references. This is the new way to browse for tags and folders references properly without any duplicates. Defaults to 'false'").Default(false))
+	Field(service.NewBoolField("browseHierarchicalReferences").Description("Set to true to browse hierarchical references. This is the new way to browse for tags and folders references properly without any duplicates. Defaults to 'false'").Default(false)).
+	Field(service.NewIntField("pollRate").Description("The rate in milliseconds at which to poll the OPC UA server when not using subscriptions. Defaults to 1000ms (1 second).").Default(1000))
 
 func ParseNodeIDs(incomingNodes []string) []*ua.NodeID {
 
@@ -123,6 +124,11 @@ func newOPCUAInput(conf *service.ParsedConfig, mgr *service.Resources) (service.
 		return nil, err
 	}
 
+	pollRate, err := conf.FieldInt("pollRate")
+	if err != nil {
+		return nil, err
+	}
+
 	// fail if no nodeIDs are provided
 	if len(nodeIDs) == 0 {
 		return nil, errors.New("no nodeIDs provided")
@@ -148,6 +154,7 @@ func newOPCUAInput(conf *service.ParsedConfig, mgr *service.Resources) (service.
 		LastMessageReceived:          atomic.Uint32{},
 		HeartbeatManualSubscribed:    false,
 		HeartbeatNodeId:              ua.NewNumericNodeID(0, 2258), // 2258 is the nodeID for CurrentTime, only in tests this is different
+		PollRate:                     pollRate,
 	}
 
 	return service.AutoRetryNacksBatched(m), nil
@@ -190,6 +197,7 @@ type OPCUAInput struct {
 	Subscription                 *opcua.Subscription
 	ServerInfo                   ServerInfo
 	BrowseHierarchicalReferences bool
+	PollRate                     int
 }
 
 // Connect establishes a connection to the OPC UA server.
