+++ opcua_plugin/opcua_unittest_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	"github.com/gopcua/opcua/id"
 	"github.com/gopcua/opcua/ua"
@@ -56,6 +57,7 @@ var _ = Describe("Unit Tests", func() {
 
 		var (
 			ctx                          context.Context
+			cncl                         context.CancelFunc
 			nodeBrowser                  NodeBrowser
 			path                         string
 			level                        int
@@ -67,7 +69,7 @@ var _ = Describe("Unit Tests", func() {
 			browseHierarchicalReferences bool
 		)
 		BeforeEach(func() {
-			ctx = context.Background()
+			ctx, cncl = context.WithTimeout(context.Background(), 2*time.Second)
 			path = ""
 			level = 0
 			logger = &MockLogger{}
@@ -76,6 +78,9 @@ var _ = Describe("Unit Tests", func() {
 			errChan = make(chan error, 100)
 			wg = &TrackedWaitGroup{}
 			browseHierarchicalReferences = false
+		})
+		AfterEach(func() {
+			cncl()
 		})
 
 		Context("When browsing nodes with a node class value nil", func() {
